Rename package to metricsreporterjmx
Both metrics-reporter-jmx and metrics-reporter-prometheus plugins used
the same package name: com.googlesource.gerrit.plugins.metricsreporters
Adding a class with the same name in both plugins would be no problem
for the bazel build but it is a problem for the generated eclipse
project. Eclipse reports an error because of a type defined multiple
times.

Change-Id: I01866284ded303550f39b5239a5dd7d2b34d6f6c