Update patch set 4
Patch Set 4: Verified+1

I suggest to go ahead and submit this series.

Here are the reasons:
Reloading of this plugin was always problematic and this issue is unrelated to this change.
As I already wrote in a previous comment, the issue is caused by the fact that:
* new plugin instance cannot register mbeans because names are already occupied by the registrations
done from the old instance
* the old plugin instance will unregister all mbeans it registered

This means that on first restart basically all mbeans are removed. However, on the next
restart the mbeans will be registered again because now there are no registered mbeans and
the new plugin instance can register them. The old plugin instance has nothing to unregister
because it previously failed to register mbeans.

The above descried behaviour/issue existed since the first version of this plugin.
I tested the version of this plugin with my changes, without my changes and
the "old" version before configuration was introduced by Mihaly. All versions have
issues with the plugin reload.

Since my series do not make anything worse than it already was and since they fix some
issues, let's submit them.

Patch-set: 4
Reviewer: Gerrit User 1003873 <1003873@173816e5-2b9a-37c3-8a2e-48639d4f1153>
Label: Verified=+1